MAJOR UP! Increased color limits, alt palettes..
MAJOR: (1) Before, you could select a color limit between 2-8.  The
palette was filled with colors that were rather 'too old-school for
comfort.'  This palette is still here, BUT... there is now a 'sort of
retro' mode, as well as a modern looking mode, BUT.. it gets better!!
(2) You can now set the color limit between 2-8, 16, 32, and even 64!!!!
You can achieve a HUGE amount of different styles with what is available
right now.  When combining these palettes, higher color limits, and your
favorite LUTs, you are in for a world of INFINITE customization!!!!!!
ENJOY!!!!!!